Code Refactoring. Added Docs to website
