+++ docs/reference/index.rst
@@ -32,3 +32,4 @@ by MIL. These subsystems relate to a variety of processes.
     pneumatic
     sabertooth
     bagging
+    preflight
