+++ docs/reference/preflight.md
@@ -1,0 +1,2 @@
+```{include} ../../mil_common/utils/mil_tools/scripts/mil-preflight/preflight.md
+```
