+++ mil_common/utils/mil_tools/scripts/mil-preflight/preflight.md
@@ -4,7 +4,6 @@ Simply type "preflight" anywhere in the MIL directory. Make sure that a robot is
 ## Description
 Preflight is an automated testing tool that should be run after turning on and connecting to the robot to run a prelaunch hardware checklist and automated software checklist.
 
-In the current version 1.0, Run Specific Test is not yet implemented.
 ### There are three types of automated software tests
 #### Actuators
 This test will prompt the user to enable physical moving actuators on the robot. Make sure that the area is cleared and the robot won't damage itself or others nearby. The user will have to watch and validate that they move as expected personally.
@@ -19,5 +18,7 @@ ROS Topics act as a channel for ROS Nodes to communicate by publishing and subsc
 
 To add a Topic test, add a topic to the list in tests.py
 
-### Dependencies
-ROS, Rich, PyInquirer
+### Setup Tests
+There are also setup tests. These are used to verify certain features on the robot that cannot be automated. For example ensuring that the O-rings are greased.
+
+To add a Setup test, add a what need to be tested to the list in tests.py
