--- services.txt
@@ -1,165 +1,0 @@
-/adaptive_controller/get_loggers
-/adaptive_controller/set_logger_level
-/adaptive_controller/set_parameters
-/alarm/get
-/alarm/set
-/alarm_sever/get_loggers
-/alarm_sever/set_logger_level
-/b_matrix
-/c3_trajectory_generator/get_loggers
-/c3_trajectory_generator/set_logger_level
-/camera/down/down_image_proc/get_loggers
-/camera/down/down_image_proc/set_logger_level
-/camera/down/down_image_proc_debayer/set_parameters
-/camera/down/down_image_proc_rectify_color/set_parameters
-/camera/down/down_image_proc_rectify_mono/set_parameters
-/camera/down/image_color/compressed/set_parameters
-/camera/down/image_color/compressedDepth/set_parameters
-/camera/down/image_color/theora/set_parameters
-/camera/down/image_mono/compressed/set_parameters
-/camera/down/image_mono/compressedDepth/set_parameters
-/camera/down/image_mono/theora/set_parameters
-/camera/down/image_raw/compressed/set_parameters
-/camera/down/image_raw/compressedDepth/set_parameters
-/camera/down/image_raw/theora/set_parameters
-/camera/down/image_rect/compressed/set_parameters
-/camera/down/image_rect/compressedDepth/set_parameters
-/camera/down/image_rect/theora/set_parameters
-/camera/down/image_rect_color/compressed/set_parameters
-/camera/down/image_rect_color/compressedDepth/set_parameters
-/camera/down/image_rect_color/theora/set_parameters
-/camera/down/set_camera_info
-/camera/down/set_parameters
-/camera/front/camera_nodelet_manager/get_loggers
-/camera/front/camera_nodelet_manager/list
-/camera/front/camera_nodelet_manager/load_nodelet
-/camera/front/camera_nodelet_manager/set_logger_level
-/camera/front/camera_nodelet_manager/unload_nodelet
-/camera/front/left/image_color/compressed/set_parameters
-/camera/front/left/image_color/compressedDepth/set_parameters
-/camera/front/left/image_color/theora/set_parameters
-/camera/front/left/image_mono/compressed/set_parameters
-/camera/front/left/image_mono/compressedDepth/set_parameters
-/camera/front/left/image_mono/theora/set_parameters
-/camera/front/left/image_raw/compressed/set_parameters
-/camera/front/left/image_raw/compressedDepth/set_parameters
-/camera/front/left/image_raw/theora/set_parameters
-/camera/front/left/image_rect/compressed/set_parameters
-/camera/front/left/image_rect/compressedDepth/set_parameters
-/camera/front/left/image_rect/theora/set_parameters
-/camera/front/left/image_rect_color/compressed/set_parameters
-/camera/front/left/image_rect_color/compressedDepth/set_parameters
-/camera/front/left/image_rect_color/theora/set_parameters
-/camera/front/left/seecam_image_proc/get_loggers
-/camera/front/left/seecam_image_proc/set_logger_level
-/camera/front/left/seecam_image_proc_debayer/set_parameters
-/camera/front/left/seecam_image_proc_rectify_color/set_parameters
-/camera/front/left/seecam_image_proc_rectify_mono/set_parameters
-/camera/front/left/set_camera_info
-/camera/front/left/set_parameters
-/camera/front/right/image_color/compressed/set_parameters
-/camera/front/right/image_color/compressedDepth/set_parameters
-/camera/front/right/image_color/theora/set_parameters
-/camera/front/right/image_mono/compressed/set_parameters
-/camera/front/right/image_mono/compressedDepth/set_parameters
-/camera/front/right/image_mono/theora/set_parameters
-/camera/front/right/image_raw/compressed/set_parameters
-/camera/front/right/image_raw/compressedDepth/set_parameters
-/camera/front/right/image_raw/theora/set_parameters
-/camera/front/right/image_rect/compressed/set_parameters
-/camera/front/right/image_rect/compressedDepth/set_parameters
-/camera/front/right/image_rect/theora/set_parameters
-/camera/front/right/image_rect_color/compressed/set_parameters
-/camera/front/right/image_rect_color/compressedDepth/set_parameters
-/camera/front/right/image_rect_color/theora/set_parameters
-/camera/front/right/set_camera_info
-/camera/front/right/set_parameters
-/camera/front/right_image_proc_debayer/get_loggers
-/camera/front/right_image_proc_debayer/set_logger_level
-/camera/front/right_image_proc_debayer/set_parameters
-/camera/front/right_image_proc_rect/get_loggers
-/camera/front/right_image_proc_rect/set_logger_level
-/camera/front/right_image_proc_rect/set_parameters
-/camera/front/right_image_proc_rect_color/get_loggers
-/camera/front/right_image_proc_rect_color/set_logger_level
-/camera/front/right_image_proc_rect_color/set_parameters
-/gazebo/apply_body_wrench
-/gazebo/apply_joint_effort
-/gazebo/clear_body_wrenches
-/gazebo/clear_joint_forces
-/gazebo/delete_light
-/gazebo/delete_model
-/gazebo/get_joint_properties
-/gazebo/get_light_properties
-/gazebo/get_link_properties
-/gazebo/get_link_state
-/gazebo/get_loggers
-/gazebo/get_model_properties
-/gazebo/get_model_state
-/gazebo/get_physics_properties
-/gazebo/get_world_properties
-/gazebo/pause_physics
-/gazebo/reset_simulation
-/gazebo/reset_world
-/gazebo/set_joint_properties
-/gazebo/set_light_properties
-/gazebo/set_link_properties
-/gazebo/set_link_state
-/gazebo/set_logger_level
-/gazebo/set_model_configuration
-/gazebo/set_model_state
-/gazebo/set_parameters
-/gazebo/set_physics_properties
-/gazebo/spawn_sdf_model
-/gazebo/spawn_urdf_model
-/gazebo/unpause_physics
-/hydrophones/hydrophones_visualization/get_loggers
-/hydrophones/hydrophones_visualization/set_logger_level
-/hydrophones/ping_locator/cross_correlation_debug_enable
-/hydrophones/ping_locator/enable
-/hydrophones/ping_locator/get_loggers
-/hydrophones/ping_locator/samples_debug_enable
-/hydrophones/ping_locator/set_logger_level
-/hydrophones/triggering/enable
-/hydrophones/triggering/filter_debug_enable
-/hydrophones/triggering/filter_debug_trigger
-/hydrophones/triggering/get_loggers
-/hydrophones/triggering/reset
-/hydrophones/triggering/sample_at_trigger_debug_enable
-/hydrophones/triggering/set_logger_level
-/hydrophones/triggering/trigger_debug_enable
-/magnetometer_vis/get_loggers
-/magnetometer_vis/set_logger_level
-/mission_runner/get_loggers
-/mission_runner/refresh_missions
-/mission_runner/set_logger_level
-/odom_estimator/get_loggers
-/odom_estimator/set_ignore_magnetometer
-/odom_estimator/set_logger_level
-/odometry_to_tf/get_loggers
-/odometry_to_tf/set_logger_level
-/poi_server/add
-/poi_server/delete
-/poi_server/get_loggers
-/poi_server/move
-/poi_server/save_to_param
-/poi_server/set_logger_level
-/poi_server/tf2_frames
-/robot_state_publisher/get_loggers
-/robot_state_publisher/set_logger_level
-/rosout/get_loggers
-/rosout/set_logger_level
-/set_mobo_kill
-/set_valve
-/simulate_go
-/simulate_hard_kill
-/simulate_soft_kill
-/tf_republisher/get_loggers
-/tf_republisher/set_logger_level
-/thruster_mapper/get_loggers
-/thruster_mapper/set_logger_level
-/transform_odometry/get_loggers
-/transform_odometry/set_logger_level
-/update_thruster_layout
-/usb_to_can_driver/get_loggers
-/usb_to_can_driver/set_logger_level
